requirements.txt
asttokens==2.4.1
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
cmake==3.25.0
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fsspec==2024.3.1
huggingface-hub==0.22.0
idna==2.10
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
lit==15.0.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytorchcv==0.0.67
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.25.1
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
timm==0.9.16
torch==2.0.1+cu118
torchattacks==3.5.1
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
zipp==3.18.1
